Ajouté les erreurs récentes indexées par type d'événements civils
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@49095 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business/src/main/java/ch/vd/uniregctb/stats/evenements/StatsEvenementsExternesResults.java
@@ -30,15 +30,15 @@ public class StatsEvenementsExternesResults {
 		}
 	}
 
-	private final Map<EtatEvenementExterne, BigDecimal> etats;
+	private final Map<EtatEvenementExterne, Integer> etats;
 	private final List<EvenementExterneErreur> erreurs;
 
-	public StatsEvenementsExternesResults(Map<EtatEvenementExterne, BigDecimal> etats, List<EvenementExterneErreur> erreurs) {
-		this.etats = etats != null ? Collections.unmodifiableMap(etats) : Collections.<EtatEvenementExterne, BigDecimal>emptyMap();
+	public StatsEvenementsExternesResults(Map<EtatEvenementExterne, Integer> etats, List<EvenementExterneErreur> erreurs) {
+		this.etats = etats != null ? Collections.unmodifiableMap(etats) : Collections.<EtatEvenementExterne, Integer>emptyMap();
 		this.erreurs = erreurs != null ? Collections.unmodifiableList(erreurs) : Collections.<EvenementExterneErreur>emptyList();
 	}
 
-	public Map<EtatEvenementExterne, BigDecimal> getEtats() {
+	public Map<EtatEvenementExterne, Integer> getEtats() {
 		return etats;
 	}
